Fix compile time errors, caused the new Cache variant
