+++ iml-gui/crate/src/components/alert_indicator.rs
@@ -3,7 +3,7 @@ use crate::generated::css_classes::C;
 use im::HashMap;
 use iml_wire_types::{Alert, AlertSeverity};
 use seed::{prelude::*, *};
-use std::cmp::max;
+use std::{cmp::max, sync::Arc};
 
 fn get_message(alerts: &[&Alert]) -> String {
     if alerts.is_empty() {
@@ -16,14 +16,15 @@ fn get_message(alerts: &[&Alert]) -> String {
 }
 
 pub(crate) fn alert_indicator<T>(
-    alerts: &HashMap<u32, Alert>,
+    alerts: &HashMap<u32, Arc<Alert>>,
     resource_uri: &str,
     compact: bool,
     tt_placement: Placement,
 ) -> Node<T> {
     let alerts: Vec<&Alert> = alerts
         .values()
-        .filter_map(|x| match &x.affected {
+        .map(|x| &**x)
+        .filter_map(|x: &Alert| match x.affected {
             Some(xs) => xs.iter().find(|x| x == &resource_uri).map(|_| x),
             None => None,
         })
